refactor(cli): extract init command module
